pyproject.toml
[project]
name = "asom-data-project"
version = "0.1.0"
description = "Data engineering project template using ASOM framework"
requires-python = ">=3.9"
dependencies = [
    "polars>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "duckdb>=1.0",
]

[project.optional-dependencies]
snowflake = [
    "snowflake-connector-python>=3.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]
all = [
    "asom-data-project[snowflake,dev]",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*", "config*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
python_files = ["test_*.py"]
python_functions = ["test_*"]
python_classes = ["Test*"]
markers = [
    "t1_logic: Unit and logic tests (T1)",
    "t2_contract: Schema and contract tests (T2)",
    "t3_quality: Data quality tests (T3)",
    "t4_access: Access control and security tests (T4)",
    "t5_idempotency: Idempotency and re-run safety tests (T5)",
    "t6_performance: Performance and cost tests (T6)",
    "t7_observability: Observability and alerting tests (T7)",
    "t8_integration: Integration and end-to-end tests (T8)",
]

[tool.ruff]
target-version = "py39"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
